refactor: ReviewSearchDto 위치 변경, SortType 수정
