+++ src/main/java/com/team04/back/domain/review/review/repository/ReviewRepositoryCustom.kt
@@ -1,7 +1,7 @@
 package com.team04.back.domain.review.review.repository
 
-import com.team04.back.domain.review.review.dto.ReviewSearchDto
 import com.team04.back.domain.review.review.entity.Review
+import com.team04.back.standard.dto.ReviewSearchDto
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
